Remove LWJGL from benchmark pending 3.0.1 release
I will revert this commit once LWJGL makes an official 3.0.1 release.

I have logged the following ticket a few days ago asking for a timeframe
on a 3.0.1 release:

https://github.com/LWJGL/lwjgl3/issues/210

An initial attempt to keep LWJGL in the build failed on Travis:

https://travis-ci.org/lmdbjava/benchmarks/builds/149336487

It is therefore simpler to just remove LWJGL for the time being and add
it back in when 3.0.1 is released.